Develop

See merge request sesuur/car365-backend!67

+++ src/main/java/ee/taltech/cars/service/ListingService.java
@@ -1,13 +1,23 @@
 package ee.taltech.cars.service;
 
 import ee.taltech.cars.dto.ParamsDto;
+import ee.taltech.cars.exception.AccessForbiddenException;
 import ee.taltech.cars.exception.ListingNotFoundException;
 import ee.taltech.cars.models.Listing;
 import ee.taltech.cars.repository.ListingRepository;
+import ee.taltech.cars.security.UserSessionHolder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.nio.file.Files;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -25,37 +35,49 @@ public class ListingService {
         return listingRepository.findById(id).orElseThrow(ListingNotFoundException::new);
     }
 
+    public List<Listing> findByOwner(UUID id) {
+        return listingRepository.findByOwner(id);
+    }
+
     public Listing save(Listing listing) {
-        return listingRepository.save(listing);
+        if (UserSessionHolder.validateAccessByID(listing.getOwner())) {
+            return listingRepository.save(listing);
+        } else throw new AccessForbiddenException();
     }
 
     public void delete(UUID id) {
-        listingRepository.delete(findById(id));
+        if (UserSessionHolder.validateAccessByID(findById(id).getOwner())) {
+            listingRepository.delete(findById(id));
+        }
     }
 
     public Listing update(Listing listing, UUID id) {
-        Listing dbListing = findById(id);
-        dbListing = new Listing(dbListing.getId(),
-                listing.getTitle(),
-                listing.getDescription(),
-                listing.getStatus(),
-                listing.getOwner(),
-                listing.getPrice(),
-                listing.getLocation(),
-                listing.getBodyType(),
-                listing.getBrand(),
-                listing.getModel(),
-                listing.getColor(),
-                listing.getGearboxType(),
-                listing.getFuelType(),
-                listing.getDriveType(),
-                listing.getEnginePower(),
-                listing.getMileage(),
-                listing.getReleaseYear(),
-                listing.getEngineSize(),
-                listing.getTime(),
-                listing.getImages());
-        return listingRepository.save(dbListing);
+        if (UserSessionHolder.validateAccessByID(findById(id).getOwner())) {
+            Listing dbListing = findById(id);
+            dbListing = new Listing(dbListing.getId(),
+                    listing.getTitle(),
+                    listing.getDescription(),
+                    listing.getStatus(),
+                    listing.isPremium(),
+                    listing.getOwner(),
+                    listing.getPrice(),
+                    listing.getLocation(),
+                    listing.getBodyType(),
+                    listing.getBrand(),
+                    listing.getModel(),
+                    listing.getColor(),
+                    listing.getGearboxType(),
+                    listing.getFuelType(),
+                    listing.getDriveType(),
+                    listing.getEnginePower(),
+                    listing.getMileage(),
+                    listing.getReleaseYear(),
+                    listing.getEngineSize(),
+                    listing.getTime(),
+                    listing.getImages());
+            return listingRepository.save(dbListing);
+        }
+        return null;
     }
 
     public List<Listing> getLatestListings(int count) {
@@ -97,5 +119,38 @@ public class ListingService {
                 .stream()
                 .map(Listing::getBrand)
                 .collect(Collectors.toList());
+    }
+
+    public void addImage(UUID id, String path) {
+        if (UserSessionHolder.validateAccessByID(findById(id).getOwner())) {
+            Listing dbListing = findById(id);
+            dbListing.addImage(path);
+            listingRepository.save(dbListing);
+        } else throw new AccessForbiddenException();
+    }
+
+    private File convertToPng(File inputFile, File outputFile) throws IOException {
+        BufferedImage bufferedImage = ImageIO.read(inputFile);
+        ImageIO.write(bufferedImage, "png", outputFile);
+        return outputFile;
+    }
+
+    public File postListingImage(MultipartFile file, UUID id) throws IOException {
+        if (UserSessionHolder.validateAccessByID(findById(id).getOwner())) {
+            final String uploadPath = "/storage/";
+            file.getOriginalFilename();
+            File convertedFile = new File(uploadPath + file.getOriginalFilename());
+            if (Arrays.asList("png", "jpeg", "jpg").contains(convertedFile.getName().split("\\.")[1])) {
+                FileOutputStream fout = new FileOutputStream(convertedFile);
+                fout.write(file.getBytes());
+                fout.close();
+                if (!convertedFile.getName().split("\\.")[1].equals("png")) {
+                    convertToPng(convertedFile, new File(uploadPath + id + ".png"));
+                    convertedFile.delete();
+                }
+                return convertedFile;
+            }
+            return null;
+        } else throw new AccessForbiddenException();
     }
 }
